#58 Enable support for char type (#79)
* #58 Enable support for char type

Support for the char type is introduced by adding a new NumericType, JavaCharType, along with a corresponding CharLiteral.
The char type is treated as a numeric type because it adheres to Java's numeric type promotion rules.

The default value for a CharLiteral is now provided by the JavaTypeSystem via the getDefaultLiteral method.

TypeASTVisitor has been updated to handle parsing of PrimitiveType.CHAR from the JDT.

ExpressionVisitor now accepts an optional targetType parameter, representing the expected type of the returned expression.
This is particularly useful for JDT's NumberLiteral, where it's ambiguous whether a literal represents a char or an int.
Since the caller may have additional context, allowing them to specify the targetType enables the visitor to assign the correct LiSA literal accordingly.

* #58 Fix testcases, rollback type propagation in visitor

Fix byte and char litaral testcases with the new introduced logic.
Rollback of the type propagation logic in the ExpressionVisitor.

* #58 Rollback AssignmentNode visitor.
Rollback changes on AssignmentNode visitor method of ExpressionVisitor.

* Minor change

* Updated literals tests, for char we store a integer

---------

Co-authored-by: VincenzoArceri <vincenzoarceri.92@gmail.com>