Fixing workflow in pipelines
All of a sudden, this fragment causes issues in the analysis execution, specifically in pipelines. Its removal resolves the problem, but I’m not sure if this reduction is acceptable.

The issue was that all analysis concluded with following error: `class "org.eclipse.core.runtime.Plugin"'s signer information does not match signer information of other classes in the same package`